refactor SignerAlgorithm.ts and SignerAlgorithm.test.ts by breaking up into files by curve
